Ejercicio 9

+++ Ejercicio 9/php/eliminar.php
@@ -12,8 +12,21 @@ if($_SERVER['REQUEST_METHOD'] == 'DELETE'){
             $CI = $_GET["CI"];
             $response['message'] = "Eliminacion realizada con éxito";
             $response['success'] = true;
-            $sql="DELETE FROM persona WHERE $CI = CI;";
-            $resultado = mysqli_query($con,$sql);
+            if (is_numeric($CI)) {
+                $sql="select * from persona where CI=".$CI;
+                $resultado1 = mysqli_query($con,$sql);
+                if (mysqli_num_rows($resultado1) > 0) {
+               
+                $sql="DELETE FROM persona WHERE $CI = CI;";
+                $resultado = mysqli_query($con,$sql);
+                } else {
+                    $response['message'] = "This CI don't exist";
+                    $response['success'] = false;
+                }
+            } else {
+                $response['message'] = "This CI don't have only numbers";
+                $response['success'] = false;
+            }
             $existe=true;
         }
     }
